+++ test/cases/inline-asset-source-svg/webpack.config.js
@@ -19,11 +19,8 @@ module.exports = {
   module: {
     rules: [
       {
-        test: /\.(png|jpe?g|ico)$/,
-        type: 'asset/resource',
-        generator: {
-          filename: 'assets/img/[name].[hash:8][ext]',
-        },
+        test: /\.(svg)$/,
+        type: 'asset/source', // test type 'asset/source' for SVG
       },
     ],
   },