+++ test/cases/loader-option-sources-attrs-filter/webpack.config.js
@@ -66,7 +66,7 @@ module.exports = {
   module: {
     rules: [
       {
-        test: /\.(png|jpe?g|ico)/,
+        test: /\.(png|jpe?g|ico)$/,
         type: 'asset/resource',
         generator: {
           filename: 'assets/img/[name].[hash:8][ext]',