+++ test/cases/resolve-relative-paths/webpack.config.js
@@ -30,7 +30,7 @@ module.exports = {
       },
 
       {
-        test: /\.(ico|png|jp?g)/,
+        test: /\.(ico|png|jp?g)$/,
         type: 'asset/resource',
         generator: {
           filename: 'assets/img/[name].[hash:8][ext]',