+++ test/cases/resolve-styles-from-module/webpack.config.js
@@ -34,7 +34,7 @@ module.exports = {
       },
       // process images loaded in styles from node module
       {
-        test: /\.(png|jpg|jpeg|ico)/,
+        test: /\.(png|jpg|jpeg|ico)$/,
         type: 'asset/resource',
         generator: {
           filename: 'assets/img/[name].[hash:8][ext]',